+++ lib/app/features/wallet/views/pages/wallet_page/components/contacts/contacts_list.dart
@@ -3,11 +3,11 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/components/screen_offset/screen_side_offset.dart';
-import 'package:ion/app/components/skeleton/container_skeleton.dart';
 import 'package:ion/app/extensions/num.dart';
 import 'package:ion/app/features/wallet/providers/contacts_data_provider.dart';
 import 'package:ion/app/features/wallet/views/pages/wallet_page/components/contacts/contacts_list_header.dart';
 import 'package:ion/app/features/wallet/views/pages/wallet_page/components/contacts/contacts_list_item.dart';
+import 'package:ion/app/features/wallet/views/pages/wallet_page/components/contacts/contacts_list_loader.dart';
 import 'package:ion/app/router/app_routes.dart';
 import 'package:ion/app/utils/username.dart';
 
@@ -25,7 +25,6 @@ class ContactsList extends ConsumerWidget {
     return contactsDataAsyncValue.maybeWhen(
       data: (contactsData) {
         final contactsDataArray = contactsData.values.toList();
-        // Use contactsDataArray as needed
         return Column(
           children: [
             const ContactListHeader(),
@@ -60,40 +59,7 @@ class ContactsList extends ConsumerWidget {
           ],
         );
       },
-      orElse: () => Column(
-        children: [
-          ContainerSkeleton(
-            width: MediaQuery.sizeOf(context).width - ScreenSideOffset.defaultSmallMargin * 2,
-            height: 16.0.s,
-            margin: EdgeInsets.symmetric(vertical: 16.0.s),
-          ),
-          SizedBox(
-            height: ContactsListItem.height,
-            child: ListView.separated(
-              padding: EdgeInsets.symmetric(
-                horizontal: ScreenSideOffset.defaultSmallMargin,
-              ),
-              scrollDirection: Axis.horizontal,
-              itemCount: 5,
-              separatorBuilder: (BuildContext context, int index) {
-                return SizedBox(width: 12.0.s);
-              },
-              itemBuilder: (BuildContext context, int index) {
-                final size = 54.0.s;
-                return ContainerSkeleton(
-                  width: size,
-                  height: size,
-                  margin: EdgeInsets.symmetric(
-                    vertical: (ContactsListItem.height - size) / 2,
-                    horizontal: (ContactsListItem.width - size) / 2,
-                  ),
-                );
-              },
-            ),
-          ),
-          footer,
-        ],
-      ),
+      orElse: () => ContactsListLoader(footer: footer),
     );
   }
 }
